% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{$.dibbler}
\alias{$.dibbler}
\alias{dibbler_accessors}
\alias{[.dibbler}
\title{Accessors for the dibbler objects}
\usage{
\method{$}{dibbler}(x, name)

\method{[}{dibbler}(x, ...)
}
\arguments{
\item{x}{A \code{dibbler} object.}

\item{name}{The name of the slot requested.}

\item{...}{Further arguments passed to other methods.}
}
\description{
These functions allow to access the content of \code{dibbler} objects.
}
\seealso{
\code{\link{make_dibbler}} to create \code{dibbler} objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
