% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{ceiling_digits}
\alias{ceiling_digits}
\title{ceiling with digits}
\usage{
ceiling_digits(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{number of decimal places}
}
\value{
numeric vector
}
\description{
ceiling with digits
}
\examples{
ceiling_digits(2.345, 1)
}
