% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibbler.R
\name{dibbler}
\alias{dibbler}
\title{The function \code{dibbler} is used for inferring the transmission of food-borne diseases across a food distribution network using pathogen genome sequences sampled for a set of cases.}
\usage{
dibbler(x, ...)
}
\arguments{
\item{x}{a list of the class 'dibbler.input' as returned by \code{dibbler.data}.}
}
\description{
!!! This package is still under development. Do not use it without contacting the author. !!!
}
\examples{

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
\item{\code{\link{dibbler.data}}}{to prepare the input data.}
}
}

