% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{floor_digits}
\alias{floor_digits}
\title{floor with digits}
\usage{
floor_digits(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{number of decimal places}
}
\value{
numeric vector
}
\description{
floor with digits
}
\examples{
floor_digits(2.345, 1)
}
