% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{igraph2data.frame}
\alias{igraph2data.frame}
\title{Convert a directed graphs from igraph to data.frame}
\usage{
igraph2data.frame(x)
}
\arguments{
\item{x}{a \code{igraph} object.}
}
\description{
The function convert a directed graph with the igraph format into into a data.frame of edge lists (from, to).
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
Conversions in dibbler include:
\item{\code{\link{network2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2network}}}{}
\item{\code{\link{network2igraph}}}{}
}
}

