% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{igraph2network}
\alias{igraph2network}
\title{Convert a directed graphs from igraph to network}
\usage{
igraph2network(x)
}
\arguments{
\item{x}{a \code{igraph} object.}
}
\description{
The function convert a directed graph with the igraph format into an network object.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

