% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{igraph2vizNetwork}
\alias{igraph2vizNetwork}
\title{Convert a directed graphs from igraph to vizNetwork inputs}
\usage{
igraph2vizNetwork(x)
}
\arguments{
\item{x}{a \code{igraph} object.}
}
\description{
The function convert a directed graph with the igraph format into inputs for vizNetwork.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
Conversions in dibbler include:
\item{\code{\link{network2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2network}}}{}
\item{\code{\link{network2igraph}}}{}
\item{\code{\link{igraph2igraph2vizNetwork}}}{}
\item{\code{\link{network2igraph2vizNetwork}}}{}
}
}

