% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{is_empty}
\alias{is_empty}
\title{Is NULL, NA, "" or 0 length}
\usage{
is_empty(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE if object is NULL, NA, "" or length 0; logical vector for objects of length > 1
}
\description{
Is NULL, NA, "" or 0 length
}
\examples{
is_empty(c(NA, "a", ""))
is_empty(NULL)
is_empty(character())
}
