% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{is_null_or_na}
\alias{is_null_or_na}
\title{Is NULL or NA}
\usage{
is_null_or_na(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE if object is NULL or NA, logical vector for objects of length > 1
}
\description{
Is NULL or NA
}
\examples{
is_null_or_na(NA)
is_null_or_na(NULL)
is_null_or_na(c(2, NA, 5))
}
