% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dibbler.R
\name{make_dibbler}
\alias{make_dibbler}
\title{Process input data for dibbler}
\usage{
make_dibbler(net, nodes_data, from = 1L, to = 2L, id = 1L)
}
\arguments{
\item{net}{A \code{data.frame} defining the edges of the food distribution
network with at least two columns representing source and receiver nodes
(i.e., 'from' and 'to'). This network is meant to include cases, which will
be seen as terminal nodes. Labels used in these columns will be matched
against \code{nodes_data}. Other columns are optional and will be kept as
edge attributes.}

\item{nodes_data}{A \code{data.frame} containing additional information on
the nodes of the network. Typically, this will be a linelist providing data
on the reported cases.}

\item{from}{An integer or character string indicating the column in
\code{net} containing source nodes ('from' column).}

\item{to}{An integer or character string indicating the column in
\code{net} containing receiving nodes ('to' column).}

\item{id}{An integer or character string indicating the column in
\code{nodes_data} containing unique case identifiers. These will be matched
against the nodes of the network described in \code{net}.}
}
\value{
a list of data suitable for input for \code{dibbler}
}
\description{
This function combines data of a food distribution network (including cases
as terminal nodes) and additional information on the nodes (typically a
linelist) to create a \code{dibbler} object. The S3 class \code{dibbler} is
an extension of the \code{epicontacts} objects implemented in the similarly
named package.
}
\examples{

if (require(outbreaks)) {

  v_data <- s_enteritidis_pt59$graph
  n_data <- data.frame(id = names(s_enteritidis_pt59$cluster),
                       cluster = s_enteritidis_pt59$cluster)

  x <- make_dibbler(v_data, n_data)
  x
}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
