% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{max_na}
\alias{max_na}
\title{max with NA handling}
\usage{
max_na(...)
}
\arguments{
\item{...}{numeric or character vector}
}
\value{
maximum or NA if all values are NA
}
\description{
max with NA handling
}
\examples{
max_na(2, 3, NA)
max_na(NA, NA)
}
