% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_na}
\alias{min_na}
\title{min with NA handling}
\usage{
min_na(...)
}
\arguments{
\item{...}{numeric or character vector}
}
\value{
minimum or NA if all values are NA
}
\description{
min with NA handling
}
\examples{
min_na(2, 3, NA)
min_na(NA, NA)
}
