% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{network2data.frame}
\alias{network2data.frame}
\title{Convert a directed graphs from network object to data.frame}
\usage{
network2data.frame(x)
}
\arguments{
\item{x}{a \code{network} object.}
}
\description{
The function convert a directed graph with the network format into a data.frame of edge lists (from, to)
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
Conversions in dibbler include:
\item{\code{\link{network2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2network}}}{}
\item{\code{\link{network2igraph}}}{}
\item{\code{\link{igraph2igraph2visNetwork}}}{}
\item{\code{\link{network2igraph2visNetwork}}}{}
}
}

