% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{network2igraph}
\alias{network2igraph}
\title{Convert a directed graphs from network to igraph}
\usage{
network2igraph(x)
}
\arguments{
\item{x}{a \code{network} object.}
}
\description{
The function convert a directed graph with the network format into an igraph object.
}
\seealso{
\describe{
Conversions in dibbler include:
\item{\code{\link{network2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2data.frame}}}{}
\item{\code{\link{igraph2network}}}{}
\item{\code{\link{network2igraph}}}{}
\item{\code{\link{igraph2visNetwork}}}{}
\item{\code{\link{network2visNetwork}}}{}
}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
