% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibbler.pal.R
\name{dibbler.pal1}
\alias{dibbler.pal1}
\alias{dibbler.pal2}
\alias{palettes}
\title{Color palettes used in dibbler}
\usage{
dibbler.pal1(n)

dibbler.pal2(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in dibbler.
}
\examples{

plot(1:4, cex=8, pch=20, col=dibbler.pal1(4), main="dibbler.pal1")
plot(1:20, col=dibbler.pal1(20), pch=20, cex=6, main="dibbler.pal1")
plot(1:10, col=dibbler.pal2(10), pch=20, cex=6, main="dibbler.pal2")
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

