% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dibbler.data.R
\name{plot.dibbler.input}
\alias{plot.dibbler.input}
\title{Plot input data for dibbler}
\usage{
\method{plot}{dibbler.input}(x, y = NULL, cex = 1, lab.cex = 1,
  color.internal = TRUE, col.pal1 = dibbler.pal1, col.pal2 = dibbler.pal2,
  ...)
}
\arguments{
\item{x}{an input dataset as returned by \code{dibbler.data}}

\item{y}{unused argument, for compatibility with the \code{plot} generic}

\item{cex}{a size factor for the nodes of the network}

\item{lab.cex}{a size factor for the tip annotations (genetic clusters)}

\item{color.internal}{a logical indicating whether internal nodes and their descending edges should be colored}

\item{col.pal1}{a color palette to be used for the genetic clusters (tips)}

\item{col.pal2}{a color palette to be used for identifying internal nodes}

\item{...}{further arguments to be passed to \code{plot.network}}
}
\value{
the same output as \code{plot.network}
}
\description{
This function plots dibbler's input data as a network, using different symbols and colors for different types of nodes. It uses the package \code{network}.
}
\examples{

if(require(igraph) && require(network)){

## generate graph from edge list
Salmonella
g <- igraph::graph.data.frame(Salmonella$graph)

input <- dibbler.data(graph=g, group=Salmonella$cluster)
plot(input)
plot(input, cex=2, color.internal=FALSE)
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link[network]{plot.network}} in the package \code{network}.
}

