% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dibbler.R
\name{plot.dibbler}
\alias{plot.dibbler}
\title{Plot input data for dibbler}
\usage{
\method{plot}{dibbler}(x, y = NULL, ...)
}
\arguments{
\item{x}{an input dataset as returned by \code{dibbler.data}}

\item{y}{unused argument, for compatibility with the \code{plot} generic}

\item{...}{further arguments to be passed to \code{plot.network}}

\item{cex}{a size factor for the nodes of the network}

\item{lab.cex}{a size factor for the tip annotations (genetic clusters)}

\item{color.internal}{a logical indicating whether internal nodes and their descending edges should be colored}

\item{col.pal1}{a color palette to be used for the genetic clusters (tips)}

\item{col.pal2}{a color palette to be used for identifying internal nodes}
}
\description{
This function plots dibbler's input data as a network, using different symbols and colors for different types of nodes. It uses the package \code{network}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
