% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dibbler.R
\name{print.dibbler}
\alias{print.dibbler}
\title{Print method for dibbler objects}
\usage{
\method{print}{dibbler}(x, ...)
}
\arguments{
\item{x}{A \code{dibbler} object.}

\item{...}{further parameters to be passed to other methods (currently not
used)}
}
\description{
This method prints the content of \code{dibbler} objects, giving an overview
of the food distribution network and the reported cases.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
