% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{range_na}
\alias{range_na}
\title{range with NA handling}
\usage{
range_na(...)
}
\arguments{
\item{...}{numeric or character vector}
}
\value{
range or c(NA, NA) if all values are NA
}
\description{
range with NA handling
}
\examples{
range_na(2, 3, NA)
range_na(NA, NA)
}
