% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Salmonella}
\alias{Salmonella}
\alias{fakeOutbreak}
\title{This dataset is a small simulated outbreak used to illustrate outbreaker.
It was generated using \code{simOutbreak}.}
\format{An object of class \code{list} of length 2.}
\source{
Gastrointestinal Bacteria Reference Unit, Public Health England
}
\usage{
Salmonella
}
\description{
This dataset is a small simulated outbreak used to illustrate outbreaker.
It was generated using \code{simOutbreak}.
}
\author{
Thomas Inns, Philip Ashton, Tim Dallman, Roberto Vivancos
}
\keyword{datasets}

