% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate.infection}
\alias{simulate.infection}
\title{Simulation of infection over a network}
\usage{
\method{simulate}{infection}(x)
}
\arguments{
\item{x}{a graph represented as a 2-column matrix storing directed edges (from, to)}
}
\description{
Under development, do not use!!
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com}) and
Tim McMackin (\email{addyour@email.here.com})
}

