% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.dibbler.data.R
\name{vis.dibbler.input}
\alias{vis.dibbler.input}
\title{Plot input data for dibbler}
\usage{
vis.dibbler.input(x, cex = 1, lab.cex = 1, plot = TRUE, legend = TRUE,
  selector = TRUE, editor = TRUE, col.pal = dibbler.pal1,
  col.internal = "#ffbf80", ...)
}
\arguments{
\item{x}{an input dataset as returned by \code{dibbler.data}}

\item{cex}{a size factor for the nodes of the network}

\item{lab.cex}{a size factor for the tip annotations (genetic clusters)}

\item{plot}{a logical indicating whether a plot should be displayed}

\item{legend}{a logical indicating whether a legend should be added to the plot}

\item{selector}{a logical indicating whether a group selector tool should be added to the plot}

\item{editor}{a logical indicating whether an editor tool should be added to the plot}

\item{col.pal}{a color palette to be used for the genetic clusters (tips)}

\item{col.internal}{a color for internal nodes}

\item{...}{further arguments to be passed to \code{visNetwork}}
}
\value{
the same output as \code{visNetwork}
}
\description{
This function plots dibbler's input data as a network, using different symbols and colors for different types of nodes. It uses the package \code{network}.
}
\examples{

if(require(igraph) && require(network)){

## generate graph from edge list
Salmonella
g <- igraph::graph.data.frame(Salmonella$graph)

input <- dibbler.data(graph=g, group=Salmonella$cluster)

## check info returned by function
temp <- vis.dibbler.input(input, plot=FALSE)
temp

\dontrun{
## opens a browser
vis.dibbler.input(input)
}
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.
}

