% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.dibbler.data.R
\name{viz.dibbler.input}
\alias{viz.dibbler.input}
\title{Plot input data for dibbler}
\usage{
viz.dibbler.input(x, cex = 1, lab.cex = 1, color.internal = TRUE,
  col.pal1 = dibbler.pal1, col.pal2 = dibbler.pal2, ...)
}
\arguments{
\item{x}{an input dataset as returned by \code{dibbler.data}}

\item{cex}{a size factor for the nodes of the network}

\item{lab.cex}{a size factor for the tip annotations (genetic clusters)}

\item{color.internal}{a logical indicating whether internal nodes and their descending edges should be colored}

\item{col.pal1}{a color palette to be used for the genetic clusters (tips)}

\item{col.pal2}{a color palette to be used for identifying internal nodes}

\item{...}{further arguments to be passed to \code{visNetwork}}
}
\value{
the same output as \code{visNetwork}
}
\description{
This function plots dibbler's input data as a network, using different symbols and colors for different types of nodes. It uses the package \code{network}.
}
\examples{

if(require(igraph) && require(network)){

## generate graph from edge list
Salmonella
g <- igraph::graph.data.frame(Salmonella$graph)

input <- dibbler.data(graph=g, group=Salmonella$cluster)
plot(input)
plot(input, cex=2, color.internal=FALSE)
}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.
}

