% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibbler.data.R
\name{dibbler.data}
\alias{dibbler.data}
\title{Process input data for dibbler}
\usage{
dibbler.data(graph = NULL, group = NULL, data = NULL)
}
\arguments{
\item{graph}{a input for buidling a graph (see details)}

\item{group}{a factor defining genetic clusters}

\item{data}{a list containing inputs for  \code{\link{dibbler}} as returned by \code{dibbler.data}}
}
\value{
a list of data suitable for input for \code{\link{dibbler}}
}
\description{
This function ...
}
\examples{

if(require(igraph)){

## generate graph from edge list
Salmonella
g <- graph.data.frame(Salmonella$graph)

dibbler.data(graph=g, group=Salmonella$cluster)

}

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

