% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibbler.R
\name{dibbler}
\alias{dibbler}
\title{Modelling food-borne outbreaks using genetic data}
\usage{
dibbler(x = dibbler.data(), ...)
}
\arguments{
\item{x}{a list of the class 'dibbler.input' as returned by \code{\link{dibbler.data}}.}

\item{...}{further arguments to be passed to other functions}
}
\description{
The function \code{dibbler} is used for inferring the transmission of food-borne diseases across a food distribution network using pathogen genome sequences sampled for a set of cases.
}
\details{
!!! This package is still under development. Do not use it without contacting the author. !!!
}
\examples{

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
\item{\code{\link{dibbler.data}}}{to prepare the input data.}
}
}

