% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono.tree.R
\name{mono.tree}
\alias{mono.tree}
\title{Idendify monochromatic trees}
\usage{
mono.tree(x = dibbler.data())
}
\arguments{
\item{x}{a list of the class 'dibbler.data' as returned by \code{\link{dibbler.data}}.}
}
\description{
The function \code{mono.tree} identifies subtrees of a food distribution network which contain only one pathogenic lineage.
}
\details{
!!! This package is still under development. Do not use it without contacting the author. !!!
}
\examples{

if(require(igraph)){
## generate graph from edge list
Salmonella
g <- graph.data.frame(Salmonella$graph)

## run mono.tree
out <- mono.tree(dibbler.data(graph=g, group=Salmonella$cluster))
names(out)
out$trees
}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
\seealso{
\describe{
\item{\code{\link{dibbler.data}}}{to prepare the input data.}
}
}

