% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibbler.pal.R
\name{dibbler_pal1}
\alias{dibbler_pal1}
\alias{palettes}
\alias{dibbler.pal1}
\alias{dibbler.pal2}
\alias{dibbler_pal2}
\title{Color palettes used in dibbler}
\usage{
dibbler_pal1(n)

dibbler_pal2(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in dibbler.
}
\examples{

plot(1:4, cex = 8, pch = 20, col = dibbler_pal1(4),
     main = "dibbler_pal1")

plot(1:20, col = dibbler_pal1(20),
     pch = 20, cex = 6, main = "dibbler_pal1")

plot(1:10, col = dibbler_pal2(10), pch = 20,
     cex = 6, main = "dibbler_pal2")
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
