% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Salmonella}
\alias{Salmonella}
\alias{fakeOutbreak}
\title{Salmonella Enteritidis P59 outbreak}
\format{An object of class \code{list} of length 2.}
\source{
Multi-agency Salmonella Enteritidis PT59 Outbreak Control Team, chaired by Public Health England
}
\usage{
Salmonella
}
\description{
This dataset covers a food-borne outbreak of Salmonella Enteritidis PT59 investigated by Public Health England. The data includes a distribution network, and genetic clusters of bacteria isolated in contaminated patients. The data is anonymised: identifiers of the food distribution network have been hashed.
}
\author{
Thomas Inns \email{Thomas.Inns@phe.gov.uk}, Philip Ashton, Tim Dallman, Roberto Vivancos
}
\keyword{datasets}
