% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.infection.R
\name{simulate.infection}
\alias{simulate.infection}
\title{Simulation of infection over a network}
\usage{
\method{simulate}{infection}(x, n.intro = 1, proba.trans = 0.8)
}
\arguments{
\item{x}{a graph represented as a 2-column matrix storing directed edges (from, to)}

\item{n.intro}{number of introductions}

\item{proba.trans}{a probability of transmission of the infection to a descending node}
}
\description{
Under development, do not use!!
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com}) and
Tim McMackin (\email{addyour@email.here.com})
}

